Num facture for Disway facture

+++ main.py
@@ -5,8 +5,7 @@ import extraction
 import contours
 from os import listdir
 import os
-from multiprocessing.dummy import Pool as ThreadPool
-import glob
+from barcode import decode
 
 images = []
 for i in range(4,5):
@@ -18,9 +17,9 @@ def is_inverse(extracted_data):
     e = extracted_data
     return 'if' not in e or 'cnss' not in e or 'patente' not in e or 'rc' not in e
 
-def extract(filename,inverse=False,resize=False,facture=False):
-    preprocessing.run(filename,inverse,resize,facture)
-    processing.run(filename,facture)
+def extract(filename,inverse=False,resize=False):
+    preprocessing.run(filename,inverse,resize)
+    processing.run(filename)
     postprocessing.run(filename)
     return extraction.run(filename)
 
@@ -34,9 +33,6 @@ def combine_extracted(extracted_data1,extracted_data2):
 
 def process(filename):
     extracted_data = extract(filename)
-    if 'facture' not in extracted_data:
-        extracted_data2 = extract(filename,False,False,True)
-        extracted_data = combine_extracted(extracted_data,extracted_data2)
     # If facture not found
     if 'facture' not in extracted_data:
         # Split image into multiple parts
@@ -46,25 +42,31 @@ def process(filename):
         images = listdir(file_path)
         for image in images:
             if '.jpg' in image:
-                extracted_data2 = extract(file_path+'/'+image,False,True,False)
+                extracted_data2 = extract(file_path+'/'+image,False,True)
                 extracted_data = combine_extracted(extracted_data,extracted_data2)
-                os.remove(file_path+'/'+image)
+        if 'facture' not in extracted_data:
+            for image in images:
+                if '.jpg' in image:
+                    result = decode(file_path+'/'+image)
+                    if result is not None:
+                        extracted_data['facture'] = result
+                    os.remove(file_path+'/'+image)
     if is_inverse(extracted_data):
-        extracted_data2 = extract(filename,True,False,False)
+        extracted_data2 = extract(filename,True,False)
         extracted_data = combine_extracted(extracted_data,extracted_data2)
     for data in extracted_data:
         extracted_data[data] = extracted_data[data].replace('\n',' ').strip()
     return extracted_data
 
-def run(images):
+"""def run(images):
     results = []
     for image in images:
-        results.append(process(image))
-    return results
+        results.append({image:process(image)})
+    return results"""
 
-"""def run():
+def run():
     results = []
     for image in images:
         results.append(process(image))
     print(results)
-run()"""
+run()
